Refactor: Move file hashing to dedicated file
